feat: allow mixing highlight colors
This update allows users to define highlight colors that are a mix
of two other colors in the theme. By specifying the mix colors as
color names from the theme (ex "green" or "red"), the highlight
groups will adapt to match the current theme as the user switches
themes. This includes switching between light and dark themes.

@example
```lua
hl_override = {
  DiffAdd = {
    -- Set the background of diff additions to a color
    -- 50% between the green and black values defined
    -- in the theme. For light themes, this will actually
    -- be a mix between green and white colors, resulting
    -- in a light green background. Dark themes will show
    -- a dark green background.
    bg = {"green", "black", 50}
  }
}
```